third_party/WebKit/LayoutTests/fast/events/touch/gesture/pad-gesture-cancel.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../../resources/js-test.js"></script>
<script src="../../../../resources/run-after-layout-and-paint.js"></script>
<script>
    description("Tests basic use of GestureFlingCancel");

    var expectedWheelEventsOccurred = "0";
    var actualWheelEventsOccurred = 0;

    function recordWheelEvent(event)
    {
        shouldBe('event.clientX', "10");
        shouldBe('event.clientY', "11");

        // Test deliberately does not equality check wheelDeltas to not be fragile in the face of curve adjustment.
        shouldBeTrue("event.wheelDeltaX > 5");
        shouldBeTrue("event.wheelDeltaY > 5");
        actualWheelEventsOccurred++;
    }

    document.addEventListener("mousewheel", recordWheelEvent);

    if (window.testRunner && window.eventSender && window.eventSender.gestureFlingStart) {
        eventSender.gestureFlingStart(10, 11, 1000, 1000, "touchpad");
        eventSender.gestureFlingCancel();
        testRunner.waitUntilDone();
    }

    runAfterLayoutAndPaint(function() {
        shouldBe('actualWheelEventsOccurred', expectedWheelEventsOccurred);
        isSuccessfullyParsed();
        if (window.testRunner)
            testRunner.notifyDone();
    });
</script>
</head>
<body>
<span id="e"></span>
<span id="f"></span>
</body>
</html>
